requirements.txt
pandas
numpy
matplotlib
requests
python-dotenv
yfinance
openai
tqdm
